Add JSONEnum builder
