+++ Sources/JSONSchemaBuilder/Schema/TypeSpecific/JSONArray.swift
