+++ Sources/JSONSchemaBuilder/Schema/TypeSpecific/JSONBoolean.swift
