+++ Sources/JSONSchemaBuilder/Schema/TypeSpecific/JSONInteger.swift
