+++ Sources/JSONSchemaBuilder/Schema/TypeSpecific/JSONNull.swift
