+++ Sources/JSONSchemaBuilder/Schema/TypeSpecific/JSONNumber.swift
