+++ Sources/JSONSchemaBuilder/Schema/TypeSpecific/JSONString.swift
